# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于Go+Vue实现的openLDAP后台管理项目",
        "title": "Go Ldap Admin",
        "termsOfService": "https://github.com/eryajf/go-ldap-admin",
        "contact": {
            "name": "项目作者：二丫讲梵 、 swagger作者：南宫乘风",
            "url": "https://github.com/eryajf/go-ldap-admin",
            "email": "https://github.com/eryajf/go-ldap-admin"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8888",
    "basePath": "/api",
    "paths": {
...
        "/role/add": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "新建角色记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色管理"
                ],
                "summary": "新建角色记录",
                "parameters": [
                    {
                        "description": "添加角色记录的结构体",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.RoleAddReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseBody"
                        }
                    }
                }
            }
        },
        "/role/delete": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除角色记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色管理"
                ],
                "summary": "删除角色记录",
                "parameters": [
                    {
                        "description": "删除角色记录的结构体",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.RoleDeleteReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseBody"
                        }
                    }
                }
            }
        },
        "/role/getapilist": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取接口列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色管理"
                ],
                "summary": "获取接口列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "roleId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseBody"
                        }
                    }
                }
            }
        },
        "/role/getmenulist": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取菜单列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色管理"
                ],
                "summary": "获取菜单列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "roleId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseBody"
                        }
                    }
                }
            }
        },
        "/role/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取角色记录列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色管理"
                ],
                "summary": "获取角色记录列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseBody"
                        }
                    }
                }
            }
        },
        "/role/update": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "更新角色记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色管理"
                ],
                "summary": "更新角色记录",
                "parameters": [
                    {
                        "description": "更新角色记录的结构体",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.RoleUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseBody"
                        }
                    }
                }
            }
        },
        "/role/updateapis": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "更新接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色管理"
                ],
                "summary": "更新接口",
                "parameters": [
                    {
                        "description": "更新接口的结构体",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.RoleUpdateApisReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseBody"
                        }
                    }
                }
            }
        },
        "/role/updatemenus": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "更新菜单",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色管理"
                ],
                "summary": "更新菜单",
                "parameters": [
                    {
                        "description": "更新菜单的结构体",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.RoleUpdateMenusReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.RoleAddReq": {
            "type": "object",
            "required": [
                "keyword",
                "name"
            ],
            "properties": {
                "keyword": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 1
                },
                "name": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 1
                },
                "remark": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 0
                },
                "sort": {
                    "type": "integer",
                    "maximum": 999,
                    "minimum": 1
                },
                "status": {
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                }
            }
        },
        "request.RoleDeleteReq": {
            "type": "object",
            "required": [
                "roleIds"
            ],
            "properties": {
                "roleIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "request.RoleUpdateApisReq": {
            "type": "object",
            "required": [
                "apiIds",
                "roleId"
            ],
            "properties": {
                "apiIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "roleId": {
                    "type": "integer"
                }
            }
        },
        "request.RoleUpdateMenusReq": {
            "type": "object",
            "required": [
                "menuIds",
                "roleId"
            ],
            "properties": {
                "menuIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "roleId": {
                    "type": "integer"
                }
            }
        },
        "request.RoleUpdateReq": {
            "type": "object",
            "required": [
                "id",
                "keyword",
                "name"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "keyword": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 1
                },
                "name": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 1
                },
                "remark": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 0
                },
                "sort": {
                    "type": "integer",
                    "maximum": 999,
                    "minimum": 1
                },
                "status": {
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                }
            }
        },
...
        "response.ResponseBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}